Move docs fragments about
