+++ plugins/doc_fragments/ec2.py
@@ -1,30 +1,21 @@
 # -*- coding: utf-8 -*-
 
-# Copyright: (c) 2015, Ansible, Inc
+# (c) 2022 Red Hat Inc.
+#
+# This file is part of Ansible
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
+from .region import ModuleDocFragment as RegionFragment
+
+#
+# The amazon.aws.ec2 docs fragment has been deprecated,
+# please migrate to amazon.aws.region.modules.
+#
+
 
 class ModuleDocFragment(object):
-
-    # EC2 only documentation fragment
-    DOCUMENTATION = r'''
-options:
-    region:
-        description:
-          - The AWS region to use.
-          - For global services such as IAM, Route53 and CloudFront, I(region)
-            is ignored.
-          - The C(AWS_REGION) or C(EC2_REGION) environment variables may also
-            be used.
-          - See the Amazon AWS documentation for more information
-            U(http://docs.aws.amazon.com/general/latest/gr/rande.html#ec2_region).
-          - The C(ec2_region) alias has been deprecated and will be removed in
-            a release after 2024-12-01
-          - Support for the C(EC2_REGION) environment variable has been
-            deprecated and will be removed in a release after 2024-12-01.
-        type: str
-        aliases: [ aws_region, ec2_region ]
-'''
+    def __init__(self):
+        self.DOCUMENTATION = RegionFragment.MODULES
